+++ src/drawingarea/sectiondrawingarea.cpp
@@ -122,6 +122,14 @@ namespace PixelMaestroStudio {
 	 * @param event Event parameters.
 	 */
 	void SectionDrawingArea::mousePressEvent(QMouseEvent *event) {
+		// Sets the current Section as the active Section on left click
+		if (event->buttons() == Qt::LeftButton) {
+			Section* active_section = maestro_drawing_area_->get_maestro_control_widget()->section_control_widget_->get_active_section();
+			if (active_section != this->section_) {
+				maestro_drawing_area_->get_maestro_control_widget()->section_control_widget_->set_active_section(this->section_);
+			}
+		}
+
 		mouseMoveEvent(event);
 	}